Added system for Entities Alarming values
Added needed methods and visual representation to enable the system to check for values that are not in the good range, so that exclamation mark can be shown, and also card to change color to pale red. This is reversed if in the card is entity which values aren't out of range.